+++ Foundation/Array/Bitmap.hs
@@ -89,6 +89,8 @@ instance C.Collection Bitmap where
     elem e = Data.List.elem e . toList
     minimum = Data.List.minimum . toList . C.getNonEmpty -- TODO can shortcircuit all this massively
     maximum = Data.List.maximum . toList . C.getNonEmpty -- TODO DITTO
+    all p = Data.List.all p . toList
+    any p = Data.List.any p . toList
 instance C.Sequential Bitmap where
     take = take
     drop = drop