+++ tests/Test/Utils/Foreign.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE ExistentialQuantification #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 module Test.Utils.Foreign
     ( Ptr
     , Storable
@@ -10,16 +12,16 @@ import           Foreign.Marshal.Alloc
 import           Foreign.Ptr
 import           Foreign.Storable
 import           Foundation
-import           Prelude (head, zip)
+import           Prelude (zip)
 import           Control.Monad (forM_)
 
 import           Foundation.Foreign
 
-createPtr :: Storable e => [e] -> IO (FinalPtr e)
+createPtr :: forall e . Storable e => [e] -> IO (FinalPtr e)
 createPtr l
     | null l    = toFinalPtr nullPtr (\_ -> return ())
     | otherwise = do
-        let szElem = sizeOf (head l)
+        let szElem = sizeOf (undefined :: e)
             nbBytes = szElem * length l
         ptr <- mallocBytes nbBytes
         forM_ (zip [0..] l) $ \(o, e) -> pokeElemOff ptr o e