+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.4
+
+* Expose Sequential from Foundation
+* Export internal withPtr for optimisation
+* Export `ifThenElse`
+* Use the proper `String` type for error instead of `[Char]`
+* Add `any` and `all` to `Collection`
+
 ## 0.0.3
 
 Monad: